Fish AI
Had some help from Deepseek to make the fish move in random directions at random times. Also had Deepseek help me with getting the fish to move towards a hook within a certain radius of the hook and delete the fish when reached a second hookedRadius. That part will be changed when full game is developed.