Card Game Updates
Both a player and AI hand exist and can gain cards. Current big step is adding a turn order so that only one side can draw cards at a time. AIManager is a copy of HandManager, but will be expanded to have automation capabilities such as asking for a card, while the HandManager will be manual.